refactor(Analysis): move component to its own file
