+++ src/components/menu/Analysis/Analysis.cy.tsx
